backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "gradio_app"]

[project]
name = "stocksense-ai"
version = "0.1.0"
description = "AI-powered Indian equity research assistant"
requires-python = ">=3.12"

dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "python-dotenv>=1.0.0",
    # Scraping
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.2.0",
    # PDF extraction
    "pdfplumber>=0.11.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.4.0",
    # News
    "newsapi-python>=0.2.7",
    # Social (Twitter/X is scraped via Google, no tweepy needed)
    # Gradio UI
    "gradio>=4.36.0",
    "openai>=2.26.0",
    "langchain>=1.2.12",
    "langchain-openai>=1.1.11",
    "langgraph>=1.1.2",
    "langchain-google-genai>=4.2.1",
    "transformers>=5.3.0",
    "torch>=2.10.0",
    "supabase>=2.28.0",
    "langchain-core>=1.2.18",
    "sentence-transformers>=5.3.0",
    "langchain-huggingface>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "pandas>=2.0.0",
    "pandas-ta>=0.3.14",
    "yfinance>=0.2.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.12.0",
    "ruff>=0.3.0",
    "black>=24.0.0",
    "httpx>=0.27.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py312"]
